Renamed some files and fixed some CI errors
