+++ pkg/utils/exec/exec.go
@@ -12,7 +12,7 @@
 //    License for the specific language governing permissions and limitations
 //    under the License.
 
-package execute
+package exec
 
 import (
 	"os/exec"
